backend/requirements.txt
fastapi==0.115.12
pydantic==2.11.3
gspread==6.2.0
openai==1.75.0
google-auth==2.39.0
google-auth-oauthlib==1.2.1
python-dotenv==1.1.0
uvicorn==0.34.2
